Made updates per initial pull request comments
