--- sensors/positioning/bno085_i2c/src/main/java/com/sample/impl/sensor/BNO085/Activator.java
@@ -1,22 +1,0 @@
-/***************************** BEGIN LICENSE BLOCK ***************************
- The contents of this file are subject to the Mozilla Public License, v. 2.0.
- If a copy of the MPL was not distributed with this file, You can obtain one
- at http://mozilla.org/MPL/2.0/.
-
- Software distributed under the License is distributed on an "AS IS" basis,
- WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
- for the specific language governing rights and limitations under the License.
-
- Copyright (C) 2020-2025 Botts Innovative Research, Inc. All Rights Reserved.
- ******************************* END LICENSE BLOCK ***************************/
-package com.sample.impl.sensor.BNO085;
-
-import org.sensorhub.utils.OshBundleActivator;
-
-/**
- * The presence of this class tells the OpenSensorHub OSGI machinery about this module.
- * It is referenced in the 'Bundle-Activator' attribute in the OSGI section of the build.gradle file.
- */
-@SuppressWarnings("unused")
-public class Activator extends OshBundleActivator {
-}
