--- sensors/positioning/bno085_i2c/src/main/resources/META-INF/services/org.sensorhub.api.module.IModuleProvider
@@ -1,1 +1,0 @@
-com.sample.impl.sensor.BNO085.Descriptor
