+++ sensors/positioning/sensorhub-driver-bno085/src/main/resources/META-INF/services/org.sensorhub.api.module.IModuleProvider
@@ -1,0 +1,1 @@
+org.sensorhub.impl.sensor.BNO085.Descriptor
